Input source refactoring. WndProc is now replaced with WindowInputSource.

+++ include/win32/input/PointerListener.hpp
@@ -8,7 +8,7 @@
 
 namespace input {
 
-typedef int pointer_button;
+typedef unsigned short pointer_button;
 typedef int pointer_id;
 
 class PointerListener
